// projects/reusable-ui-library/src/lib/paginator/paginator.interface.ts
/**
 * structure for the paginator instance
 */
export interface PaginatorInstance {

    /** id of the pagination instance */
    id: string;
  
    /** total numeber of items to be paginated. usually array length */
    totalItems: number;
  
    /** current page number */
    currentPage: number;
  
    /** total items to be shown per page */
    itemsPerPage: number;
  }
  
  /**
   * structure for each page item
   */
  export interface Page {
  
    /** number of the page */
    number: number;
  
    /** text to be displayed for the page numeber */
    text: string;
  
    /** whether the page is active or not */
    active: boolean;
  }
  
  export interface PaginatorPipeArgs {
    /** id of the pagination instance */
    id?: string;
  
    /** current page number, defaults to 1 */
    currentPage?: number;
  
    /** total items to be shown per page, defaults to 10 */
    itemsPerPage?: number;
  }
  